1.2/requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.5.0
pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
streamlit>=1.25.0
tqdm>=4.60.0
pandas>=1.3.0
requests>=2.25.0